### 2024-10-15 #7355 @editorialbot
[PRE REVIEW]: aweSOM: a CPU/GPU-accelerated Self-organizing Map and Statistically Combined Ensemble Framework for Machine-learning Clustering Analysis
**Submitting author:** @tvh0021 (<a href="http://orcid.org/0000-0001-6600-2517">Trung Ha</a>)
**Repository:** https://github.com/tvh0021/aweSOM
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
