### 2024-11-26 #7355 @stsievert
@gkthiruvathukal sorry, I can't review. I don't have the bandwidth now.

I have taken a brief glance. I'd encourage the authors (@tvh0021, ...) to review [the JOSS review guidelines][1]. In particular, I think the main page should feature a clear statement of need and the answer to the question "what's the purpose of this project?" An FAQ page is not where I'd expect to find that answer.

[1]:https://joss.readthedocs.io/en/latest/review_criteria.html
