### 2024-12-03 d4d0b20
add a statement of need to the front page of the documentation
