+++ docs/source/index.rst
@@ -20,7 +20,7 @@ statistically combined ensemble (SCE) using `ensemble learning <https://github.c
 .. aweSOM is developed specifically to identify intermittent structures (current sheets) in 3D plasma simulations 
 .. (`link to arXiv pre-print <https://doi.org/10.48550/arXiv.2410.01878>`_). However, it can also be used for a variety of clustering and classification tasks.
 
-aweSOM was developed for machine-learning clustering and classification tasks, with a focus on identifying intermittent structures in 3D plasma simulations.
+aweSOM was developed for machine-learning clustering and classification tasks, with a focus on identifying intermittent structures in 3D plasma simulations (`link to arXiv pre-print <https://doi.org/10.48550/arXiv.2410.01878>`_).
 Past implementations of SOM and SCE were single-threaded and not optimized for large datasets, so aweSOM was developed to address this issue.
 Using a combination of JIT-accelerated and parallelized SOM and GPU-accelerated SCE, aweSOM can handle datasets with up to :math:`\sim 10^8` points running on a single GPU/CPU node.
 
