### 2024-12-03 #7355 @tvh0021
@stsievert thank you for your feedback!
@gkthiruvathukal, I have addressed this issue by including a brief statement of need on the front page of the [documentation](https://awesom.readthedocs.io/en/latest/). A complete statement of need for the project is available in the manuscript.
