### 2024-12-08 #7355 @haniyeka
> @haniyeka, @sherbold, @stsievert: Please let me know if you are willing to contribute a review for this JOSS submission. I need at least 2 reviewers to proceed.

Thanks for the invitation. I do have some time available to review this submission. Feel free to add me as reviewer.
