### 2024-12-30 tvh0021/aweSOM#1 @apizzuto
JOSS Article Proof clarification question (part of https://github.com/openjournals/joss-reviews/issues/7613)
I've begun my review of your JOSS submission, and started with the paper. I had one question regarding Figure 1, specifically where you mention in the Figure caption

```
The dotted lines shows linear extrapolations from the data in order to estimate the speedup
```

I didn't quite follow this sentence. I did not think the dotted lines on the left were extrapolations, I thought they represented the raw runtimes from POPSOM tests. Could you please clarify?
