### 2024-12-31 tvh0021/aweSOM#1 @tvh0021
Thank you for your question! I was not being clear in the caption. The extrapolated dotted line refers to the extended olive line in the top-right panel, for the SCE analysis. The dashed lines in the left panel are all raw runtime of POPSOM. I will make an update to the caption to clarify this point.
