### 2024-12-31 84c50a5
Clarify paper caption
