+++ src/aweSOM/combine_clumps.py
@@ -1,0 +1,1 @@
+## Given segmentational labels, combine clumps of the same label that are connected
