### 2025-01-07 f81901f
update with clump finding algorithm
