+++ src/aweSOM/combine_clumps.py
@@ -1,1 +1,0 @@
-## Given segmentational labels, combine clumps of the same label that are connected
