### 2025-01-17 tvh0021/aweSOM#2 @apizzuto
Verifying performance claims in JOSS manuscript
Hey aweSOM authors!

I'm wrapping up my review for https://github.com/openjournals/joss-reviews/issues/7613#issuecomment-2558050818, and in your article proof, there are performance claims relative to the current state of the art. As part of review, we are asked about verifying these claims:

```
Performance: If there are any performance claims of the software, have they been confirmed?
```

Is there any way you could add something to your tutorials so that someone new to aweSOM could do benchmarking? It isn't necessary to replicate every part of that plot, necessarily, but would be great to have the user input an N and F value and then run and profile how long this took in aweSOM. I don't think it would be necessary to include doing this for the legacy SOM or SCE software, because I understand that would add way too much complexity for installations, etc., I just want to make sure we could replicate the rough aweSOM numbers
