### 2025-01-17 tvh0021/aweSOM#3 @apizzuto
Suggestion for iris tutorial notebook
Hello authors!

Just a suggestions for your tutorial notebook, but this is not a blocker for my review in https://github.com/openjournals/joss-reviews/issues/7613#issuecomment-2558050818.

I found it strange that in the jupyter notebook, you were sending users to the command line to run commands and then come back to the notebook. Would there be any way to expose some function in `sce.py` to do this so that you didn't have to run the script?
