### 2025-01-17 #7613 @apizzuto
Hi authors! Thank you for your submission of a wonderful piece of software. I enjoyed running the tutorials, I found your API docs very thorough and complete, and I checked your pytest coverage and was impressed.

I just had a couple of issues that I submitted to your repo. The only one that I think is a blocker for my review is regarding verifying performance / runtime claims in the proof: https://github.com/tvh0021/aweSOM/issues/2. The other issue (https://github.com/tvh0021/aweSOM/issues/3) is just a suggestion, but otherwise, from my perspective, my review is complete. I'll mark my remaining checklist items once the performance claims issue is resolved
