### 2025-01-17 tvh0021/aweSOM#2 @tvh0021
Hi @apizzuto,

Thank you for your comment! Users can in fact verify the performance claims of both the SOM and SCE parts of aweSOM, following the tutorial in the  [performance test](https://awesom.readthedocs.io/en/latest/testing.html#performance-tests) section.

Specifically, to benchmark the performance of the aweSOM software given values of `N` and `F`, you can run `python popsom_bench.py --N [N] --F [F] --awesom` inside the `aweSOM/tests/performance/` directory. The `--awesom` flag tells the script to only benchmark aweSOM, not the legacy POPSOM code.

For example, the script outputs: `[{'N': 10000, 'F': 5, 'N_nodes': 171, 'aweSOM': {'train': 5.039193153381348, 'map': 4.368703842163086}}]` and `[{'N': 1000000, 'F': 5, 'N_nodes': 1830, 'aweSOM': {'train': 168.09887099266052, 'map': 8.321974039077759}}]`, which specifies N, F, and number of nodes in the lattice. It also lists the time taken, in seconds, to train and map the SOM to the data.

A similar benchmark can be run for the SCE code using the `sce_bench.py` script. The data created is synthetic, and do not correspond to any real-world application.

Please let me know if this addresses your feedback and if you would like me to provide any additional information!
