+++ examples/plasma-turbulence/combine_clumps.py
@@ -1,7 +1,44 @@
 ## Given segmentation labels, combine clumps of the same label that are connected
 
 import numpy as np
-from scipy.ndimage import label
+from tqdm import tqdm
+
+
+def generate_neighbors(connectivity):
+    """Generate relative neighbor positions based on connectivity. If data is 3D, assume z-axis is the first axis."""
+    if connectivity == 4:  # 2D direct neighbors only
+        deltas = [(0, -1), (-1, 0), (1, 0), (0, 1)]
+    elif connectivity == 8:  # 2D
+        deltas = [(-1, -1), (-1, 0), (-1, 1), (0, -1), (0, 1), (1, -1), (1, 0), (1, 1)]
+    elif connectivity == 6:  # 3D direct neighbors only
+        deltas = [
+            (dx, dy, dz)
+            for dx in [-1, 0, 1]
+            for dy in [-1, 0, 1]
+            for dz in [-1, 0, 1]
+            if abs(dx) + abs(dy) + abs(dz) == 1
+        ]
+    elif connectivity == 10:  # 3D, 8-connectivity in 2D plus +z and -z
+        deltas = [
+            (dx, dy, dz)
+            for dx in [-1, 0, 1]
+            for dy in [-1, 0, 1]
+            for dz in [-1, 0, 1]
+            if abs(dx) + abs(dy) + abs(dz) == 1
+        ] + [(0, -1, -1), (0, 1, -1), (0, -1, 1), (0, 1, 1)]
+    elif connectivity == 26:  # 3D
+        deltas = [
+            (dx, dy, dz)
+            for dx in [-1, 0, 1]
+            for dy in [-1, 0, 1]
+            for dz in [-1, 0, 1]
+            if not (dx == 0 and dy == 0 and dz == 0)
+        ]
+    else:
+        raise ValueError(
+            "Unsupported connectivity. For 2D, use 4 or 8; for 3D, use 6, 10, or 26."
+        )
+    return deltas
 
 
 def get_neighbors(
@@ -68,9 +105,67 @@ def get_clumps(data_array: np.ndarray, connectivity: np.ndarray) -> list:
                             clumps.append(clump)
     else:
         raise ValueError("Input array must be 2D or 3D.")
-        return None
 
     return clumps
+
+
+def merge_double_sheets(
+    double_sheet_dict: dict, positive_sheet_dict: dict, negative_sheet_dict: dict
+):
+    merged_double_sheets = dict()
+    neighbor_deltas = generate_neighbors(10)
+
+    for double_key, double_positions in double_sheet_dict.items():
+        merged_clump = set()
+        print("Double sheet:", double_key)
+
+        has_positive_neighbor = False
+        has_negative_neighbor = False
+
+        visited_positive_keys = set()
+        visited_negative_keys = set()
+
+        for double_position in tqdm(double_positions):
+            # print('Double position:', double_position)
+            neighbors = [
+                tuple(np.add(double_position, delta)) for delta in neighbor_deltas
+            ]
+
+            # check positive neighbors
+            for positive_key, positive_positions in positive_sheet_dict.items():
+                if any([neighbor in positive_positions for neighbor in neighbors]):
+                    has_positive_neighbor = True
+                    merged_clump.update(positive_positions)
+                    visited_positive_keys.add(positive_key)
+                    break
+
+            if has_positive_neighbor:
+                # check negative neighbors
+                for negative_key, negative_positions in negative_sheet_dict.items():
+                    if any([neighbor in negative_positions for neighbor in neighbors]):
+                        has_negative_neighbor = True
+                        merged_clump.update(negative_positions)
+                        visited_negative_keys.add(negative_key)
+                        break
+
+            # break condition stops the loop as soon as one positive and one negative neighbor is found
[…]
