### 2025-01-18 887cd55
added joss badge in documentation
