### 2025-01-18 dd35c42
remove badge from documentation
