### 2025-01-18 tvh0021/aweSOM#3 @tvh0021
Hi @apizzuto,

Thank you for your suggestion! I have edited the iris.ipynb file and updated the [documentation](https://awesom.readthedocs.io/en/latest/notebooks/iris.html#Stacking-multiple-aweSOM-realizations-for-SCE-analysis) to reflect this change.

I had to use `%run -i` to pass through command line instructions inside a Jupyter Notebook. There are probably more sophisticated approaches to achieve this, but since the Iris dataset is so small the SCE script can be run on pretty much any Jupyter environment in minimal time.
