### 2025-01-21 tvh0021/aweSOM#3 @apizzuto
Hey @tvh0021 – thanks for taking a look at this. I think your change does fix the main thing I was thinking about, but I think when I initially suggested I was thinking of putting a function in `sce.py` (e.g. `def run_sce()`) where the args are equivalent to the args you'd pass via command line and so you could call `sce.run_sce` as a function directly from the notebook, instead of having to run something that would be a command line command.

Regardless, I think the flow is much nicer now, because you can do everything from the notebook, so feel free to close this out!
