### 2025-01-22 tvh0021/aweSOM#3 @tvh0021
Ah you are right, that would make it a lot easier. It did not cross my mind to do so.

I will implement this change in an upcoming update. Thanks again for your suggestions!
