### 2025-02-28 #7613 @Abinashbunty
@gkthiruvathukal Apologies on my side for the super delay in responding. I checked that I had somehow unsubscribed from repository for notifications. I had indeed completed my review few weeks ago but couldn't mark the items as checked nor drop a comment here with updates.

Let me know if you can still assign or consider the review. I went through full execution of the documentation. I didn't find anything that should be changed. The code owner has done a remarkable job in this paper and its documentation. I had some improvement points, but found that resolved in further changes that were done while I was reviewing. 💯

Apart from that, a general note: Awesome package just like that name and very well written! I'm sure there will be a lot of users who will come across this and will find it very useful, especially those in academics at the moment.

Again, apologies for my delay in responding! I adjusted my GitHub settings already for future. 🙏🏻
