### 2025-03-02 5c2260c
fix doi references
