+++ joss/paper.bib
@@ -191,7 +191,8 @@ The final program has been completed, and it is necessary to guarantee the corre
   title        = {{Iris}},
   year         = {1936},
   howpublished = {UCI Machine Learning Repository},
-  note         = {{DOI}: https://doi.org/10.24432/C56C76}
+  note         = {{DOI}: https://doi.org/10.24432/C56C76},
+  doi		   = {10.24432/C56C76}
 }
 
 @software{pandas,
@@ -221,7 +222,7 @@ The final program has been completed, and it is necessary to guarantee the corre
 
 @article{Pei2023,
 	author = {Dehao Pei and Chao Luo and Xiaomei Liu},
-	doi = {https://doi.org/10.1016/j.asoc.2022.109972},
+	doi = {10.1016/j.asoc.2022.109972},
 	issn = {1568-4946},
 	journal = {Applied Soft Computing},
 	keywords = {Time series, Self-organizing map, Financial trading system, Unsupervised learning},
@@ -236,7 +237,7 @@ The final program has been completed, and it is necessary to guarantee the corre
 
 @article{Alvarez2008,
 	author = {Manuel Alvarez-Guerra and Cristina Gonz{\'a}lez-Pi{\~n}uela and Ana Andr{\'e}s and Berta Gal{\'a}n and Javier R. Viguri},
-	doi = {https://doi.org/10.1016/j.envint.2008.01.006},
+	doi = {10.1016/j.envint.2008.01.006},
 	issn = {0160-4120},
 	journal = {Environment International},
 	keywords = {Sediment, Quality assessment, Self-Organizing Map, Multivariate statistical techniques},
