### 2025-03-27 ba3cc1b
Minor documentation and capability update, post-review version for JOSS publication
