### 2025-03-27 #7613 @tvh0021
Hi @gkthiruvathukal, thank you for providing me with the checklist. I will check them off as I complete them and will ping you when all are completed.

> * [x]  Double check authors and affiliations (including ORCIDs)
> * []  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [ ]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [ ]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [ ]  Make sure that the license listed for the archive is the same as the software license.
>
> For your convenience, I also organized it as a task list.
