### 2025-03-28 #7613 @tvh0021
Hi @gkthiruvathukal, I have completed the checklist above.

The version that will be used for the JOSS paper: v1.1.0

The DOI of the Zenodo archive: https://doi.org/10.5281/zenodo.15098625

Thank you!
